FIX 알림기능도 가능 김기훈 2025-02-15 18:27

+++ Prepay/app/src/main/java/com/example/prepay/data/response/OrderHistoryRes.kt
@@ -2,7 +2,7 @@ package com.example.prepay.data.response
 
 data class OrderHistoryRes(
     val orderHistoryId : Int,
-    val orderDate: Int,
+    val orderDate: Long,
     val totalPrice: Int,
     val refundRequested: Boolean,
     val storeImgUrl: String,
